xcm_runtime_apis::authorized_aliases

Trait AuthorizedAliasersApi

pub trait AuthorizedAliasersApi<Block: BlockT>: Core<Block> {
    // Provided methods
    fn authorized_aliasers(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        target: VersionedLocation,
    ) -> Result<Result<Vec<OriginAliaser>, Error>, ApiError> { ... }
    fn is_authorized_alias(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        origin: VersionedLocation,
        target: VersionedLocation,
    ) -> Result<Result<bool, Error>, ApiError> { ... }
}

API for querying XCM authorized aliases

Provided Methods

fn authorized_aliasers( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, target: VersionedLocation, ) -> Result<Result<Vec<OriginAliaser>, Error>, ApiError>

Returns locations allowed to alias into and act as target.

fn is_authorized_alias( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, origin: VersionedLocation, target: VersionedLocation, ) -> Result<Result<bool, Error>, ApiError>

Returns whether origin is allowed to alias into and act as target.