Module xcm_runtime_apis::conversions

Exposes runtime APIs for various XCM-related conversions. Contains runtime APIs for useful conversions, such as between XCM Location and AccountId.

Structs

LocationToAccountHelper

A helper implementation that can be used for LocationToAccountApi implementations. It is useful when you already have a ConvertLocation<AccountId> implementation and a default Ss58Prefix.

Enums

Error

Traits

LocationToAccountApi

API for useful conversions between XCM Location and AccountId.