xcm_runtime_apis

Module dry_run

Dry-run API. Given an extrinsic or an XCM program, it returns the outcome of its execution. Runtime API definition for dry-running XCM-related extrinsics. This API can be used to simulate XCMs and, for example, find the fees that need to be paid.

Structs

CallDryRunEffects

Effects of dry-running an extrinsic.

XcmDryRunEffects

Effects of dry-running an XCM program.

Enums

Error

Traits

DryRunApi

API for dry-running extrinsics and XCM programs to get the programs that need to be passed to the fees API.