xcm_runtime_apis

Module trusted_query

Exposes runtime API for querying whether a Location is trusted as a reserve or teleporter for a given Asset. Runtime API definition for checking if given <Asset, Location> is trusted reserve or teleporter.

Enums

Error

Traits

TrustedQueryApi

API for querying trusted reserves and trusted teleporters.

Type Aliases

XcmTrustedQueryResult

Result of TrustedQueryApi functions.